Create new class ItemAdapaterContextMenuStation
